refactored automl as one module in sam_ml
