+++ docs/class_generator.py
@@ -4,11 +4,12 @@ from typing import Callable
 
 from jinja2 import Environment, FileSystemLoader
 
+import sam_ml.automl
 import sam_ml.data.preprocessing
 import sam_ml.models.classifier
 import sam_ml.models.regressor
+from sam_ml.automl.main_auto_ml import AutoML
 from sam_ml.data.preprocessing.main_data import Data
-from sam_ml.models.main_auto_ml import AutoML
 from sam_ml.models.main_classifier import Classifier
 from sam_ml.models.main_model import Model
 from sam_ml.models.main_regressor import Regressor
