+++ sam_ml/__init__.py
@@ -3,4 +3,5 @@ __version__ = "1.2.1"
 __all__ = [
     "models",
     "data",
+    "automl",
 ]
