+++ sam_ml/automl/__init__.py
@@ -1,0 +1,8 @@
+from .ClassifierTest import CTest
+from .RegressorTest import RTest
+
+__all__ = [
+    # Non-modules:
+    "CTest",
+    "RTest",
+]
