+++ sam_ml/models/regressor/__init__.py
@@ -4,12 +4,10 @@ from .ElasticNet import EN
 from .ExtraTreesRegressor import ETR
 from .LassoLarsCV import LLCV
 from .RandomForestRegressor import RFR
-from .RegressorTest import RTest
 from .SGDRegressor import SGDR
 from .XGBoostRegressor import XGBR
 
 __all__ = [
-    "RTest",
     "RFR",
     "DTR",
     "ETR",
